feat(components): expose lightweight join surface
Extract the workspace join request flow from its router adapter so hosted entries can compose it without importing the full route tree or workspace runtime.

Model: gpt-5
